… in a way that dependencies that are not required by another unsatisfied dependency are resolved first

Alternative: #8256

# Pull Request Check List

Resolves: (partially<sup>1</sup>) boto3/botocore vs. urllib3 issue
Closes: #8256

<!-- This is just a reminder about the most common mistakes. Please make sure that you tick all *appropriate* boxes.  But please read our [contribution guide](https://python-poetry.org/docs/contributing/) at least once, it will save you unnecessary review cycles! -->

- [x] Added **tests** for changed code.
- [ ] Updated **documentation** for changed code.

<!-- If you have *any* questions to *any* of the points above, just **submit and ask**!  This checklist is here to *help* you, not to deter you from contributing! -->

Currently, we choose to resolve dependencies with less possible versions first. We can improve the heuristics, which dependency to resolve first, by considering relations between dependencies:

If dependency `A` depends on dependency `B`, we should resolve dependency `A` first no matter which dependency has more versions. Of course, different versions of a package can have different dependencies, but most often dependencies are at least similar and don't change much between versions. So I think the new heuristics will not worsen the performance in most cases (and of course improve it in some cases) - at least if the heuristics itself can be calculated fast enough.

Example: If latest `A` requires `B<2`, then older versions of `A` will typically not allow newer versions of `B` and if they do it's only because the incompatibility had not been known at the time they were released.

Some measurements (with warm cache):

|`pyproject.toml` from ...| time without PR | time with PR |
|---|---|---|
|shootout example|225 s|8 s|
|shootout example with additional `urllib3<2` constraint|8 s|8 s|
|#4670|1 s|1 s|
|#4870|70 s|118 s|
|large internal project|64 s|52 s|

As can be seen, the performance of the shootout example improves dramatically. The performance of the #4870 example is worse, however, we also get some different (probably better) results:

|#4870 without PR|#4870 with PR|
|---|---|
|Sphinx 3.5.3 (requires docutils>=0.12)<br/>docutils 0.18.1|Sphinx 4.3.2 (requires docutils>=0.14,<0.18)<br/>docutils 0.17.1|

The solution with the PR is probably better because an older Sphinx version probably does not work with a newer docutils version. The constraint is just missing because the incompatibility was not known when the older Sphinx version was released. Since that's a typical issue, **the new heuristics may even lead to less surprising results for inexperienced users if there are several possible solutions.**

<sup>1</sup> This PR can solve the boto3/botocore vs. urllib3 issue in some cases (like the shootout example), but not in all cases. That's because boto3 does not depend directly on urllib3 but on botocore, which depends on urllib3. With this PR we only look forward one level. That means it doesn't help if we have to choose between boto3 and urllib3 with no other dependency that depends on urllib3 in the list of unsatisfied dependencies yet as in the example in #7950 with boto3 and urllib3 as the only top level dependencies.