@dimbleby 2023-07-30
Have you thought about the approach being explored in #8191, which simply reverses the preference for "packages with fewer versions"?

That is tempting to me in that it is a single character fix - just add a minus sign.  And presumably it will also handle the cases described in your footnote.

I suppose all heuristics are in the end subject to bad cases - I'm pretty sure that the problem ought to be NP-hard.  While I see the logic in trying to force conflicts sooner rather than later: perhaps that heuristic is optimising for a case that turns out to be not very important in the python ecosystem, whereas "packages with most versions" would be helpful in the world as it actually is.

My experience is that: in most solutions most packages are at the latest allowed version anyway.  So simply eliminating the known pathological case might be a pretty sensible try.