@radoering 2023-07-30
No, I haven't really thought about it. I just assumed that there is a good reason for choosing dependencies with less versions first - there's even a test (`test_traverse_into_package_with_fewer_versions_first`). On the other hand, the code is quite old so that we won't find out if there were some real-world examples that benefit from that decision.

Choosing packages with more versions first, of course, solves the boto3/urllib3 issue completely and also the Sphinx/docutils example. Testing the projects from the description, I can measure a slight performance regression for the shootout example with the additional urllib constraint: When choosing packages with less versions first it takes 8 s, when choosing packages with more versions first it takes 10 s. Since we have no example yet where it's worse, maybe, we should risk it...