@dimbleby 2023-07-30
the original algorithm describes "fewest versions first" - https://github.com/dart-lang/pub/blob/master/doc/solver.md#decision-making

> Pub chooses the latest matching version of the package with the fewest versions that match the outstanding constraint. This tends to find conflicts earlier if any exist, since these packages will run out of versions to try more quickly. But there's likely room for improvement in these heuristics.

I suspect the testcase was written just because that's how the algorithm is described.  While that text acknowledges room for improvement, it would be amusing if that improvement was doing the exact opposite!

The way I'm thinking about it is that - in cases where there are real conflicts to resolve - reversing this heuristic probably loses some average performance, but helps with worst-case performance

- most of the time it makes sense to try and find dead-ends sooner rather than later, per the original heuristic
- but we have this known expensive case where poetry reaches a dead-end (eg with urllib3==2.0) but can't tell that it's a dead end without working its way through a zillion versions of eg botocore, and there it helps a _lot_ to do things the other way round

But also most of the time there just aren't that many conflicts anyway - as I say, mostly most things end up at the latest version - so my hope is that the damage this does to the average case isn't much to worry about.

However I've a feeling that I could tell myself a story justifying almost any heuristic!  Perhaps there's no way really to know except to ship it and see what the new bad cases are...